==> S3_GO_barplot.R <==
#DavidHCollins
#6 08 2020

#This script is relevant to the following publication:

#David H. Collins1*, Anders Wirén1, 2*, Marjorie Labédan1, 3, Michael Smith1, David C. Prince1, Irina Mohorianu1, 4, Tamas Dalmay1, and Andrew F. G. Bourke1. Gene expression in larval caste determination of bumblebees and the transition to advanced eusociality.

#This script is to use the GO analysis data generated by Gorilla and Revigo to produce a barplot that summarises the  GO results.

#File required: 
#'All_GO_output' - Contains the GO output for each phenotype

###############################################################################################

#First you must alwys reset R
rm(list=ls())

#make these packages (and associated functions) available to use in the script

library(readr)
library(tidyverse)
library(gridExtra)
library(ggstance)
library(grid)


#Set working directory
setwd("C:/Users/dhcollins500/OneDrive - University of East Anglia/Documents/Academic writing/My papers/Collins et al_BB-M001482-1_obj1 paper/R scripts/")

all <- read_csv("All_GO_output.csv")


#Remove GO terms eliminated by Revigo
df <- all %>% 
  filter(Eliminated==0)


#Make a new column called sig
df <- df %>% 
  mutate(sig=-log10(P_value))

#Reorder factor levels in Phenotype, Development, and Caste
df$Development <- factor(df$Development,levels = c("Early (EQ)", "Mid (MQ)", "Late (LQ)", "Early (EW)", "Mid (MW)", "Late (LW)"))
df$Phenotype <- factor(df$Phenotype,levels = c("EQ","EW","MQ","MW","LQ","LW"))
df$Caste <- factor(df$Caste,levels = c("queen", "worker"))


#Due to huge number of GO terms and confusing information, limit to just the Biological Processes terms
df3 <- df %>% 
  filter(Ontology=="Process")

#These data could be presented in three different ways: 1) one figure with everything, but that limits the number of GO terms that can be displayed; 2) two figures of queen and workers with, early, mid, late on each, but two figures is still quite limiting and the three developmental phenotypes on one figure doesn't make the most sense; 3) three figures of early, mid and late, with queen and worker seperated on each figure, but three figures is a lot of figures, and does go together on a page very well. As such, seperating two figures for the two castes seems the least worst option.


#Due to huge number of GO terms, I will limit the number than can appear on any one graph to 20
df3 <- df3 %>% 
  group_by(Phenotype) %>% 
  slice_max(order_by = desc(P_value), n = 20)


#Explore numbers
df3 %>% 
  group_by(Caste) %>% 
  select(Phenotype) %>% 
  tally()



#Make an order variable that is each phenotype with each ontology in descending order of p-values ranked (to make it easier to arrange figures)
df3 <- df3 %>% 
  group_by (Phenotype, Development) %>% 
  arrange(desc(P_value),.by_group = TRUE) %>% 
  ungroup(Development) %>% 
  mutate (order = row_number())


### Important : When you try to reorder within a facet, because of the way the faceting works it messes up the order somewhat ###

### Therefore create the following functions to produce a new 'reorder_within' function that reorders the groups within each facet (in this  case phenotype)

reorder_within <- function(x, by, within, fun = mean, sep = "___", ...) {
  new_x <- paste(x, within, sep = sep)
  stats::reorder(new_x, by, FUN = fun)
}


#' @rdname reorder_within
#' @export
scale_x_reordered <- function(..., sep = "___") {
  reg <- paste0(sep, ".+$")
  ggplot2::scale_x_discrete(labels = function(x) gsub(reg, "", x), ...)
}


#' @rdname reorder_within
#' @export
scale_y_reordered <- function(..., sep = "___") {
  reg <- paste0(sep, ".+$")
  ggplot2::scale_y_discrete(labels = function(x) gsub(reg, "", x), ...)
}

#Note the new functions replace the "reorder" and "scale_x_discrete" functions. In all other respects they should act the same as these two standard functions in ggplot.

# Rename the facets



(queen_bp <- df3%>% 
    filter(Caste=="queen") %>% 
    ggplot(aes(y=reorder_within(Description,order,Phenotype),x=-log10(P_value), fill=Development)) +
    geom_barh(stat="identity", colour="black", width=.7)+
    scale_y_reordered(name ="Gene ontology term")+
    scale_x_continuous(name ="-log10(P value)")+
    scale_fill_manual(values = c("#F0F0F0","#636363","#BDBDBD"))+ 
    facet_grid(Development~., scales = "free_y", space = "free_y") +
    labs(title = "A")+
    theme(plot.title.position = "plot", 
          text = element_text(size=12),
          axis.text.x = element_text(),
          axis.line= element_line(),
          legend.position = "none",
          panel.border = element_blank(), 
          panel.background = element_blank(), 
          panel.grid = element_blank(), 
          panel.spacing.x = unit(0.3,"line")) )

#Use this function to confirm which colours were used. Originally the palette was "scale_colour_brewer(palette = "greys")", I then swapped which greys appeared to get the colours in the order that I wanted (light - dark - medium). This was to make the figure consistent with other figures in the paper.

g <- ggplot_build(queen_bp)
unique(g$data[[1]]["fill"])

(worker_bp <- df3%>% 
    filter(Caste=="worker") %>% 
    ggplot(aes(y=reorder_within(Description,order,Phenotype),x=-log10(P_value), fill=Development)) +
    geom_barh(stat="identity", colour="black", width=.7)+
    scale_y_reordered(name ="Gene ontology term")+
    scale_x_continuous(name ="-log10(P value)")+
    scale_fill_manual(values = c("#F0F0F0","#636363","#BDBDBD"))+ 
    facet_grid(Development~., scales = "free_y", space = "free_y") +
    labs(title = "B")+
    theme(plot.title.position = "plot", 
          text = element_text(size=12),
          axis.text.x = element_text(),
          axis.line= element_line(),
          legend.position = "none",
          panel.border = element_blank(), 
          panel.background = element_blank(), 
          panel.grid = element_blank(), 
          panel.spacing.x = unit(0.3,"line")) )



ggsave("GOfigure_queen.pdf", queen_bp, width = 7, height = 7)

ggsave("GOfigure_worker.pdf", worker_bp, width = 7, height = 7)
